fixed some refactoring issues
