+++ basic-electron-react-boilerplate/src/assets/about.png
