+++ basic-electron-react-boilerplate/src/assets/background.png
