+++ basic-electron-react-boilerplate/src/assets/cambutton.png
