+++ basic-electron-react-boilerplate/src/assets/camera-off.png
