+++ basic-electron-react-boilerplate/src/assets/gestures/fist.png
