+++ basic-electron-react-boilerplate/src/assets/gestures/four-fingers.png
