+++ basic-electron-react-boilerplate/src/assets/gestures/one-finger.png
