+++ basic-electron-react-boilerplate/src/assets/gestures/open-hand.png
