+++ basic-electron-react-boilerplate/src/assets/gestures/peace.png
