+++ basic-electron-react-boilerplate/src/assets/gestures/rock-and-roll.png
