+++ basic-electron-react-boilerplate/src/assets/gestures/three-fingers.png
