+++ basic-electron-react-boilerplate/src/assets/home.png
