+++ basic-electron-react-boilerplate/src/assets/logo.png
