+++ basic-electron-react-boilerplate/src/assets/profile.png
