+++ basic-electron-react-boilerplate/src/assets/settings.png
