+++ basic-electron-react-boilerplate/src/assets/webcam.png
