+++ basic-electron-react-boilerplate/src/assets/zoom in.png
