+++ basic-electron-react-boilerplate/src/components/Main/Main.js
