+++ basic-electron-react-boilerplate/src/components/MainScreen/CameraOption.js
