+++ basic-electron-react-boilerplate/src/components/MainScreen/CameraScreen.js
