+++ basic-electron-react-boilerplate/src/components/MainScreen/MainLayout.js
