+++ basic-electron-react-boilerplate/src/components/MainScreen/Options.js
