+++ basic-electron-react-boilerplate/src/components/Settings/GestureBox.js
