+++ basic-electron-react-boilerplate/src/components/Settings/HandButtons.js
