+++ basic-electron-react-boilerplate/src/components/Settings/SettingsLayout.js
