+++ basic-electron-react-boilerplate/src/components/Settings/TrickGesturePair.js
