+++ basic-electron-react-boilerplate/src/components/Settings/TrickGesturePairs.js
