+++ front-end/src/App.js
@@ -1,6 +1,4 @@
 import './App.css';
-import MainLayout from "./components/MainScreen/MainLayout";
-import SettingsLayout from "./components/Settings/SettingsLayout"
 import Main from "./components/Main/Main"
 
 function App() {
@@ -10,7 +8,7 @@ function App() {
       <Main />
     </div>
   );
-  
+
 }
 
 export default App;
